+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Fixed text sometimes missing from game options section headers
 - Fixed faction selection issue with recent BAR versions
 - Disable maximized starting window states, including cli option, to avoid bug where combo-box options show up on top left corner
+- Modified main window to show a banner, skylobby and engine discord links and identify the fork name 
 
 ## [0.9.31008](https://github.com/springraaar/skylobby/releases/tag/0.9.31008) - 2025-10-24